fix: add models to events, moveconstants to properties; test: refactor test
