+++ audit-lib-spring-boot-starter/src/main/java/io/github/auditlib/http/IncomingHttpFilter.java
@@ -1,0 +1,69 @@
+package io.github.auditlib.http;
+
+import io.github.auditlib.model.HttpAuditEvent;
+import io.github.auditlib.service.AuditLogServiceImpl;
+import jakarta.servlet.Filter;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.stereotype.Component;
+import org.springframework.web.util.ContentCachingRequestWrapper;
+import org.springframework.web.util.ContentCachingResponseWrapper;
+
+import java.io.IOException;
+
+/**
+ * Фильтр для логирования входящих HTTP запросов
+ */
+@Component
+public class IncomingHttpFilter implements Filter {
+
+    private final AuditLogServiceImpl auditLogService;
+
+    public IncomingHttpFilter(AuditLogServiceImpl auditLogService) {
+        this.auditLogService = auditLogService;
+    }
+
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
+            throws IOException, ServletException {
+
+        HttpServletRequest httpRequest = (HttpServletRequest) request;
+        HttpServletResponse httpResponse = (HttpServletResponse) response;
+
+        ContentCachingRequestWrapper wrappedRequest = new ContentCachingRequestWrapper(httpRequest);
+        ContentCachingResponseWrapper wrappedResponse = new ContentCachingResponseWrapper(httpResponse);
+
+        try {
+
+            chain.doFilter(wrappedRequest, wrappedResponse);
+
+        } finally {
+
+            logHttpRequest(wrappedRequest, wrappedResponse);
+
+            wrappedResponse.copyBodyToResponse();
+
+        }
+    }
+
+    private void logHttpRequest(ContentCachingRequestWrapper request, ContentCachingResponseWrapper response) {
+
+        String requestBody = new String(request.getContentAsByteArray());
+        String responseBody = new String(response.getContentAsByteArray());
+
+        HttpAuditEvent event = HttpAuditEvent.createIncomingEvent(
+                request.getMethod(),
+                request.getRequestURI(),
+                requestBody,
+                response.getStatus(),
+                responseBody
+        );
+
+        auditLogService.logHttpEvent(event);
+    }
+
+}
