+++ audit-lib-spring-boot-starter/src/main/java/io/github/auditlib/service/KafkaService.java
@@ -1,0 +1,12 @@
+package io.github.auditlib.service;
+
+import io.github.auditlib.model.HttpAuditEvent;
+import io.github.auditlib.model.MethodAuditEvent;
+
+public interface KafkaService {
+
+     boolean sendMethodEventToKafka(MethodAuditEvent event);
+
+     boolean sendHttpEventToKafka(HttpAuditEvent event);
+
+}
